+++ sample/Benchmark.java
@@ -1,9 +1,9 @@
 import java.io.File;
 import java.io.IOException;
 import java.net.InetAddress;
+import java.util.Map;
 import java.util.Random;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import com.maxmind.db.CHMCache;
 import com.maxmind.db.InvalidDatabaseException;
 import com.maxmind.db.NoCache;
@@ -45,7 +45,7 @@ public class Benchmark {
         for (int i = 0; i < count; i++) {
             random.nextBytes(address);
             InetAddress ip = InetAddress.getByAddress(address);
-            JsonNode t = r.get(ip);
+            Map t = r.get(ip, Map.class);
             if (TRACE) {
                 if (i % 50000 == 0) {
                     System.out.println(i + " " + ip);