+++ src/main/java/com/maxmind/db/DatabaseRecord.java
@@ -1,14 +1,13 @@
 package com.maxmind.db;
 
-import com.fasterxml.jackson.databind.JsonNode;
-
 import java.net.InetAddress;
 
 /**
- * Record represents the data and metadata associated with a database lookup.
+ * DatabaseRecord represents the data and metadata associated with a database
+ * lookup.
  */
-public final class Record {
-    private final JsonNode data;
+public final class DatabaseRecord<T> {
+    private final T data;
     private final Network network;
 
     /**
@@ -18,7 +17,7 @@ public final class Record {
      * @param ipAddress    the IP address used in the lookup.
      * @param prefixLength the network prefix length associated with the record in the database.
      */
-    public Record(JsonNode data, InetAddress ipAddress, int prefixLength) {
+    public DatabaseRecord(T data, InetAddress ipAddress, int prefixLength) {
         this.data = data;
         this.network = new Network(ipAddress, prefixLength);
     }
@@ -27,7 +26,7 @@ public final class Record {
      * @return the data for the record in the database. The record  will be
      * <code>null</code> if there was no data for the address in the database.
      */
-    public JsonNode getData() {
+    public T getData() {
         return data;
     }