+++ src/main/java/com/maxmind/db/NodeCache.java
@@ -2,14 +2,12 @@ package com.maxmind.db;
 
 import java.io.IOException;
 
-import com.fasterxml.jackson.databind.JsonNode;
-
 public interface NodeCache {
 
     interface Loader {
-        JsonNode load(int key) throws IOException;
+        DecodedValue load(CacheKey key) throws IOException;
     }
 
-    JsonNode get(int key, Loader loader) throws IOException;
+    DecodedValue get(CacheKey key, Loader loader) throws IOException;
 
 }